requirements.txt
Flask==3.1.2
python-dotenv==1.1.1
requests==2.32.5
psycopg==3.2.9
psycopg_binary==3.2.9
psycopg_pool==3.2.6
